Text can now be rendered with the 3d scene.

+++ The WildPixels/InputHandler.cpp
@@ -74,8 +74,8 @@ void InputHandler::Update()
 	// Convert the mouse coordinates to NDC space.
 
 	// This converts the coordinates to min/max [0, 1]
-	INSTANCE->x_mouse_pos = ((double)INSTANCE->mouse->GetState().x / (double)Graphics::GetWidth());
-	INSTANCE->y_mouse_pos = ((double)INSTANCE->mouse->GetState().y / (double)Graphics::GetHeight());
+	INSTANCE->x_mouse_pos = ((float)INSTANCE->mouse->GetState().x / (float)Graphics::GetWidth());
+	INSTANCE->y_mouse_pos = ((float)INSTANCE->mouse->GetState().y / (float)Graphics::GetHeight());
 
 	// Final convert to NDC [-1, 1] with [0, 0] being in the middle of the screen.
 	// Y is positive upward and X is positive to the right.
